digi: add support to ARM64 with 64-bit memory addressing
Use type of memory addressing that depends on the platform
architecture to support ARM and ARM64

https://jira.digi.com/browse/DUB-805

+++ board/digi/common/helper.c
@@ -658,7 +658,7 @@ void verify_mac_address(char *var, char *default_mac)
  *	0 if the block is not empty
  *	-1 on error
  */
-int media_block_is_empty(u32 addr, uint hwpart)
+int media_block_is_empty(uintptr_t addr, uint hwpart)
 {
 	size_t len;
 	int ret = -1;
